requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
jinja2==3.1.3
scikit-learn==1.3.2
joblib==1.3.2
pandas==2.2.1
numpy==1.26.4
mlflow==2.11.3
pymongo==4.6.3
python-dotenv==1.0.1
email-validator==2.1.0.post1  # if email validation used
